requirements.txt
et-xmlfile==1.1.0
lxml==4.9.2
openpyxl==3.1.2
python-docx==0.8.11
xlrd==2.0.1
